Adds marketdata and utils packages
